Fix signed/unsigned comparison

+++ oscpp/client.hpp
@@ -219,7 +219,7 @@ namespace Client
         // @throw std::invalid_argument if blob size is greater than std::numeric_limits<int32_t>::max()
         Packet& blob(const Blob& arg)
         {
-            if (arg.size > std::numeric_limits<int32_t>::max())
+            if (arg.size > (size_t)std::numeric_limits<int32_t>::max())
                 throw std::invalid_argument("Blob size greater than maximum value representable by int32_t");
             m_tags.putChar('b');
             m_args.putInt32(arg.size);
